Refactor brand-related code to feature-based structure
- Created new feature directory: src/features/brand/
- Moved all brand-related components to features/brand/components/
- Moved BrandContext to features/brand/context/
- Moved brand API files to features/brand/api/
- Moved brand helpers to features/brand/utils/
- Moved BrandKnowledgePage to features/brand/pages/
- Updated all import statements across the codebase
- Maintains backward compatibility with existing functionality

This refactoring follows the feature-based structure guidelines in CLAUDE.md,
improving code organization and maintainability.
