+++ src/App.tsx
@@ -1,6 +1,6 @@
 import React, { useState, useEffect } from 'react';
 import './index.css';
-import { BrandContext, BusinessInfo } from './context/BrandContext';
+import { BrandContext, BusinessInfo } from './features/brand/context/BrandContext';
 import { CampaignContext } from './context/CampaignContext';
 import { AuthProvider } from './features/auth/context/AuthContext';
 import { withErrorBoundary } from 'react-error-boundary';
@@ -9,7 +9,7 @@ import { useAuth } from './features/auth/context/AuthContext';
 import posthog from './helpers/posthog';
 
 import './App.css';
-import BrandDataLoader from './components/BrandDataLoader';
+import BrandDataLoader from './features/brand/components/BrandDataLoader';
 import AppRoutes from './AppRoutes';
 import { AppProvider } from './context/appContext';
 import { UsersContextProvider } from './context/UsersContext';
