+++ src/components/Forms/PurposeForm.tsx
@@ -6,7 +6,7 @@ import WebsiteOwnership from '../WebsiteOwnership';
 import CampaignFormInput from '../shared/Inputs/CampaignFormInput';
 import Dropdown from '../../features/campaign/components/CampaignDropdown';
 import LanguageSelector from '../../features/campaign/components/LanguageSelector';
-import { BrandContext } from '../../context/BrandContext';
+import { BrandContext } from '../../features/brand/context/BrandContext';
 import { useAuth } from '../../features/auth/context/AuthContext';
 import posthog from '../../helpers/posthog';
