+++ src/components/PromotionComponent.tsx
@@ -3,12 +3,12 @@
 import React, { useContext, useState } from 'react';
 import { Link, useNavigate } from 'react-router-dom';
 import { CampaignContext } from '../context/CampaignContext';
-import CardStats from './dashboard/CardStats';
-import CampaignReviewButton from './dashboard/CampaignReviewButton';
+import CardStats from '../features/dashboard/components/CardStats';
+import CampaignReviewButton from '../features/dashboard/components/CampaignReviewButton';
 import handleNavigate from '../helpers/handleNavigate';
 import { useAuth } from '../features/auth/context/AuthContext';
 import { GiElectric } from 'react-icons/gi';
-import CampaignThreadWin from './campaign/CampaignThreadWin';
+import CampaignThreadWin from '../features/campaign/components/CampaignThreadWin';
 import useFetchThreadMessages from '../hooks/useFetchThreadMessages';
 import { CircularProgress } from '@mui/material';
