+++ src/features/campaign/components/StepComponentCampaign.tsx
@@ -1,6 +1,6 @@
 import React, { useContext, useState } from 'react';
 import { CampaignContext, CampaignInfoType } from '../context/CampaignContext';
-import { BrandContext } from '../context/BrandContext';
+import { BrandContext } from '../features/brand/context/BrandContext';
 import { useNavigate } from 'react-router-dom';
 // import { getAuth } from "firebase/auth";
 // import { useAuthentication } from "../firebase";
