+++ src/features/dashboard/pages/Dashboard.test.tsx
@@ -4,7 +4,7 @@ import { vi, describe, it, expect, beforeEach, afterEach } from 'vitest';
 import { BrowserRouter } from 'react-router-dom';
 import Dashboard from './Dashboard';
 import { AuthContext } from '../../auth/context/AuthContext';
-import { BrandContext } from '../../../context/BrandContext';
+import { BrandContext } from '../../brand/context/BrandContext';
 
 // Mock the modules
 vi.mock('../../../helpers/getServiceURL', () => ({
